Ask for user input on adding new lines to docs-public/build.gradle
Also update the script to use "docs(" and "samples(" as the
docs-public/guild.gradle prefix.

This script used to automatically insert new lines into
docs-public/build.gradle.  While this seems valuable, it
actually led to this mistake: aosp/1493290.  This is
because the old system required users to specifically
opt-out when creating a project instead of opting in
when importing prebuilts.

So now the script asks the user one by one if it should
add new lines to the public docs, and you have to individually
opt in.

Test: ./import_release_prebuilts.py 6983923 --groups androidx.wear --no-commit
Change-Id: Ib2c7927e1899a7773565a1acd8e3fecff9862371