Change artifactId format for import_release_prebuilts.py
1. artifactId format fix
Using just the artifact id was problematic  because
it was not possible to interpret the group id.

For example, androidx.car.app has an artifact app
but the script had no way of determining which
group id it came from.

Searching the repo was an option, but this seems
much more reasonable.

2. Add a main() to import_release_prebuilts.py
3. Add a test file for the new functions

Fixes: b/185511481
Test: ./test_import_release_prebuilts.py
Test: ./import_release_prebuilts.py 7280249 --artifacts androidx.car.app:app --no-commit
Change-Id: I57043974b539b9e691daa9bb47dd6c75cf36a7ae