Temporary workaround to build androidx.wear.compose:compose-foundation-samples
(see aosp/1769845)

androidx.wear.compose:compose-foundation-samples only needs Compose 1.0.0-rc02,
so I created this special version of the samples that depends on 1.0.0-rc02.
It was retrieved from build P24622731.

When the samples are updated, we can revert this CL back the canonical version
of this artifact.  It's only used for documentation, so I think this kind of
hand crafting is safe.

Test: ./gradlew buildOnServer
Change-Id: I4965644200174014b3fcf7f880adb7f96468f696