Update import_release_prebuilts.py commit method and message
This would create commits that reference one source
when in fact the source can now come from multiple sources.
This also fixes a breakage where you don't pass in a source
for --long-form that broke the commit_prebuilts function

Additionally, use this work to update the commit message
for each of these messages.  It's much more readable
as a list.

Test: ./import_release_prebuilts.py --long-form 7675667/androidx.activity
Test: ./import_release_prebuilts.py --source 7675667 --groups androidx.activity
Change-Id: Ia24a01c579a5dea30945b4f97798128e07785a20