Update import_release_prebuilts.py to guess the presences of -samples
When importing an artifact, this script just copies that
AAR.  However, artifact often have associated samples with
a `-samples` suffix.  These are not configured in our release
tooling, but we do want to build documentation against them.
This means we actually do want to guess the presense of these
samples and document them!

When we import groups, samples are already imported because
they are implicitly part of the group.

Test an artifact with samples...
Test: ./import_release_prebuilts.py --long-form 8281210/androidx.core:core-performance --no-commit

Test an artifact without samples...
Test: ./import_release_prebuilts.py --long-form 8281210/androidx.appcompat:appcompat,8281210/androidx.autofill:autofill --no-commit

Test both artifacts with samples and without samples...
Test: ./import_release_prebuilts.py --long-form 8281210/androidx.core:core-performance,8281210/androidx.appcompat:appcompat,8281210/androidx.autofill:autofill --no-commit

Test our unit tests...
Test: ./test_import_release_prebuilts.py

Change-Id: Icd05f0744c5e2d1e9b3237d330d176fd49c58c67