Update import_release_prebuilts.py to handle already-imported libraries
This script relies on git status to determine what new libraries
or versions are available.  This means it specifically ONLY
checked for new directories in the git status and ignored
any modified files.  However, if we already imported a prebuilt
for testing or some other reason, the script will assume that
the update is not relevant for docs.  We actually do want
the script to assume that even modified files should be treated
as new versions/libraries.

The fix is just to remove filter that ignores modified files in
the git status.

Test: ./import_release_prebuilts.py --long-form 8767956/androidx.collection --no-commit
Change-Id: I395d61cc959d0fb86de4c0f821175d52ee1210d3