Fix manually added version metadata files
I manually added version metadata files for projects that hadn't released since version metadata started being generated in aosp/2654882 and aosp/2687866. If a project's version had already been bumped past the last released version, the version metadata file might reference an unreleased version. E.g. the last released version of security is 1.1.0-alpha06, but because the version has been bumped to 1.1.0-alpha07 in source, and because we treat all APIs added in an alpha version as added in the latest alpha, the docs say that APIs were added in unreleased version 1.1.0-alpha07.

This corrects the versions in the manually added files which are still being used for docs generation.

Bug: 389908573
Test: ./gradlew docs-public:docs runs successfully
Change-Id: Iea9fc21201c77e901366ac32d1e3665a4c73d536