Update compose foundation prebuilts to remove extra () after @sample
The changes from aosp/3715566, copied to the source jars for the project.

Going forward there will be an error in dackka from extra characters being added to the sample identifier. This updates the existing prebuilts used for dackka tests (1.5.0) and the published docs (1.9.0-rc01) to avoid the error. The extra () was just ending up as extra incorrect text in the docs.

Bug: 427708573
Test: Linked dackka change shows how this updates the docs
Change-Id: I2c86deb8d42c9da73786c89bd0f4fec86950410b